Allow number or name for translator type in buildDummyTranslator()
This reverts c6b78da69d6, which changed it to expect numbers when I
noticed the type being undefined in debug output, but apparently the
only tests where the type actually mattered passed it as a name.

At some point we should just change all tests to pass as a name.