Don't close Zotero on macOS when window is closed
Keep the application open with limited menus, including a Window ->
Zotero option bound to (Cmd-0) that can be used to reopen the main
window.

This allows Zotero to use almost no CPU (instead of a few percentage
points when the window is open, even when hidden with Cmd-H, though it'd
be nice to fix that) and still allow saves from the connector.

There might be some things that aren't being destroyed properly when the
window is closed or other bugginess, but it seems to work pretty well.